added working dir changes to keep working dir unique
if simulations were run more than once per min they would write to the
same directory and overwrite eachother. this was fixed by adding a check
to see if the new directory exists and then adding a _# on the end until
it does not exist. This does NOT garentee that runing the program twice
at the same time will not result in data curruption